fix(web): unblock storybook helper lint check
