+++ docs/specs/vn2e9-wide-shell-1660/SPEC.md
@@ -188,6 +188,7 @@ None
 - 2026-04-07: 主人已确认视觉结果可继续，`gpt-5.4` review-loop 未发现可执行阻塞项；本地 `build + test + build-storybook + targeted e2e` 维持通过，PR #298 已打 `type:minor` + `channel:stable` 并收敛到 merge-ready。
 - 2026-04-07: 为清除 PR freshness gate，同步 `origin/main` 并重新生成 Storybook 宽屏截图；最新工作树在同步后再次通过 `build + test + build-storybook + targeted e2e`，视觉证据已刷新到当前待推送 head。
 - 2026-04-07: fresh review 指出页面级 Storybook surface 外层 gutter 会让 `desktop1660` 少于真实壳层宽度；已把横向 padding 收回 `app-shell-boundary` 内部，并再次重跑 `build + test + build-storybook + targeted e2e`，同时刷新宽屏截图到最新工作树。
+- 2026-04-07: 本地按 CI workflow 复现 `Lint & Format Check`，确认失败点是 `web/src/components/storybookPageHelpers.tsx` 触发 `react-refresh/only-export-components`；已把 `jsonResponse` 抽到独立 `storybookResponse.ts`，并重新通过 `cargo check --locked --all-targets --all-features`、`bun run check:bun-first`、`cd web && bun run lint`、`cd web && bun run storybook:build -- --quiet`、`cd docs-site && DOCS_BASE=/codex-vibe-monitor/ bun run build` 与 `assemble-pages-site.sh`。
 
 ## 参考（References）
